Split out redux stores into inventory and profile reducers
